[all-in-compose] Module refactor
