+++ app/src/main/java/com/litus_animae/refitted/module/ExerciseRepositoryModule.kt
@@ -1,4 +1,4 @@
-package com.litus_animae.refitted
+package com.litus_animae.refitted.module
 
 import com.litus_animae.refitted.data.ExerciseRepository
 import com.litus_animae.refitted.data.room.RoomDynamoExerciseRepository
