+++ app/src/main/java/com/litus_animae/refitted/module/RefittedRoomModule.kt
@@ -1,4 +1,4 @@
-package com.litus_animae.refitted
+package com.litus_animae.refitted.module
 
 import android.content.Context
 import com.litus_animae.refitted.data.WorkoutPlanNetworkService
