+++ app/src/main/java/com/litus_animae/refitted/module/SavedStateRepositoryModule.kt
@@ -1,7 +1,9 @@
-package com.litus_animae.refitted
+package com.litus_animae.refitted.module
 
+import com.litus_animae.refitted.data.SavedStateRepository
 import com.litus_animae.refitted.data.WorkoutPlanRepository
 import com.litus_animae.refitted.data.room.RoomCacheWorkoutPlanRepository
+import com.litus_animae.refitted.data.room.RoomSavedStateRepository
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -9,10 +11,10 @@ import dagger.hilt.android.components.ViewModelComponent
 
 @Module
 @InstallIn(ViewModelComponent::class)
-abstract class WorkoutPlanRepositoryModule {
+abstract class SavedStateRepositoryModule {
 
     @Binds
     abstract fun bindRepository(
-        repositoryImpl: RoomCacheWorkoutPlanRepository
-    ): WorkoutPlanRepository
+        repositoryImpl: RoomSavedStateRepository
+    ): SavedStateRepository
 }
