+++ app/src/main/java/com/litus_animae/refitted/module/WorkoutPlanRepositoryModule.kt
@@ -1,9 +1,7 @@
-package com.litus_animae.refitted
+package com.litus_animae.refitted.module
 
-import com.litus_animae.refitted.data.SavedStateRepository
 import com.litus_animae.refitted.data.WorkoutPlanRepository
 import com.litus_animae.refitted.data.room.RoomCacheWorkoutPlanRepository
-import com.litus_animae.refitted.data.room.RoomSavedStateRepository
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -11,10 +9,10 @@ import dagger.hilt.android.components.ViewModelComponent
 
 @Module
 @InstallIn(ViewModelComponent::class)
-abstract class SavedStateRepositoryModule {
+abstract class WorkoutPlanRepositoryModule {
 
     @Binds
     abstract fun bindRepository(
-        repositoryImpl: RoomSavedStateRepository
-    ): SavedStateRepository
+        repositoryImpl: RoomCacheWorkoutPlanRepository
+    ): WorkoutPlanRepository
 }
